+++ documentation/docs/release-notes.md
@@ -5,6 +5,8 @@ title: Release notes&#58;
 
 **v3.6.0**:
 
+- Multiple authn context class refs can be set in the SAML protocol support
+
 **v3.5.0**:
 
 - Added `UserInfoOidcAuthenticator` to authenticate a user based on an access token received from an OpenID Connect login process